Improve ECG filtering in ecg.ecg (#12)
* Improve ECG filtering in ecg.ecg

The current bandpass filter in ecg.ecg has cutoff frequencies between 3 and 45 Hz. The 3 Hz cutoff significantly alters or even removes some important features from the signal (such as the P and T waves). A FIR filter with a low-pass cutoff frequency of 0.67 Hz and order equal to 1.5*sampling_rate seems to effectively remove the low-frequency components while not being too computationally expensive.

* Update ecg.py

* Add comment

To indicate the removal of the DC offset.

* Update README.md

---------

Co-authored-by: Rafael Silva <rafael.c.silva@tecnico.ulisboa.pt>